requirements.txt
azure-ai-textanalytics
python-dotenv
requests